requirements.txt
pandas==2.2.2
tqdm==4.66.4
tensorboard==2.18.0
torch==2.5.1
torchvision==0.20.1
hydra-core==1.3.2
omegaconf==2.3.0
scikit-learn==1.5.2
opencv-python==4.10.0.84
protobuf==3.20.3
torchmetrics==1.6.1